Azure_Function/requirements.txt
azure-functions
azure-ai-vision-imageanalysis>=1.0.0b3
pyodbc